removing node 4
Since the semver push is happening all over node, the build is breaking with node 4 on the default npm. Peer dependencies are getting mucked up everywhere. Node 5 is ok because npm 3 does dependencies differently. So until the semver updates across npm packages are complete, we will shut off this build on travis.